requirements.txt
streamlit
psycopg2-binary
openai
numpy
scikit-learn
plotly
scipy
kneed
flask
gunicorn